fix(sticky): demo拆解与规范 (#2024)
* fix(sticky): demo拆解与规范

* fix: 修复sticky微信小程序demo显示异常
