+++ src/packages/sticky/demos/taro/demo2.tsx
@@ -1,0 +1,13 @@
+import React from 'react'
+import { Button, Sticky } from '@nutui/nutui-react-taro'
+
+const Demo2 = () => {
+  return (
+    <>
+      <Sticky threshold={120}>
+        <Button type="primary">距离顶部120px</Button>
+      </Sticky>
+    </>
+  )
+}
+export default Demo2
